Refactor process state into its own package
Creates a new lib/auth/state package that now contains all items
related to process identity and local storage. Since these things
are consumed by client tools isolating them in their own package
will allow tools to trim the dependency tree by shedding lib/auth
for a smaller package. No functional changes have been added in
addition to moving code around. There are aliases to the old
types left around to prevent breaking builds. They will be removed
in a follow up when teleport.e is updated to consume the new package.
