+++ lib/auth/state/state_unix.go
@@ -19,7 +19,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package auth
+package state
 
 import (
 	"context"
